Add error codes, recovery suggestions, and doc comments to DeviceSigner module
